Show the same sub-block again by default
Analysing the code in the main menu I realised the current sub-block was
being displayed again unless an action explicitly moved to the next
block or sub-block. I moved that common code from the if
block processing the result of `menu_listdups` to the end of the loop,
so it will be executed unless the loop is broken.