Sort indexes to delete in exec_delete
Using the indexes without sorting meant some undesired behavioir.
Removing elements of the list changes the indexes, so if an lower index
is removed before a higher one, the element in the higher one is not the
expected.

To properly sort the tuple, the values have to be ints. For str, "111"
preceeds "2".