Fix infinite loop in menu_select_dups
An infinite loop occurred when the user selected many options to delete
as well as CANCEL or ALL. These options take priority over all others so
the user is prompted if they want to execute them or no. Upon selecting
no, the option was removed but the loop displaying the menu was not
exited, resulting in an infinite loop and eventually raising a KeyError.

This commit breaks the loop after removing cacelling these options.